% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lasvmPredict.R
\name{lasvmPredict}
\alias{lasvmPredict}
\title{lasvmPredict}
\usage{
lasvmPredict(x, model, verbose = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{model}{trained model}

\item{verbose}{verbose output?}
}
\value{
a list consisting of
	predictions		the predicted labels
}
\description{
Use lasvm to train a given problem.
}

