% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R, R/lasvmTrain.R
\name{lasvmTrain}
\alias{lasvmTrain}
\title{lasvmTrain}
\usage{
lasvmTrain(x, y, gamma, cost, degree = 3, coef0 = 0, optimizer = 1,
  kernel = 2, selection = 0, termination = 0, cachesize = 256,
  bias = 1, epochs = 1, epsilon = 0.001, verbose = FALSE)

lasvmTrain(x, y, gamma, cost, degree = 3, coef0 = 0, optimizer = 1,
  kernel = 2, selection = 0, termination = 0, cachesize = 256,
  bias = 1, epochs = 1, epsilon = 0.001, verbose = FALSE)
}
\arguments{
\item{x}{data matrix}

\item{y}{labels}

\item{degree}{degree for poly kernel}

\item{coef0}{coefficient for poly kernel}

\item{optimizer}{type of optimizer}

\item{kernel}{kernel type}

\item{selection}{selection strategy}

\item{termination}{criterion for stopping}

\item{cachesize}{size of kernel cache}

\item{bias}{use  bias?}

\item{epochs}{number of epochs}

\item{epsilon}{stopping criterion parameter}

\item{verbose}{verbose output?}

\item{rounds}{number of rounds (orthogonal views)}

\item{k}{number of clusters}

\item{iter}{numer of iterations in one round}

\item{initType}{centroid initialization via Random or KMeans++}

\item{x}{data matrix}

\item{verbose}{verbose output?}
}
\value{
a list consisting of
	centers	these are the resulting centroids of the kmean algorithm (as a std::vector of NumericMatrix)
	cluster 	these are the labels for the resulting clustering (as a std::vector of NumericVector)
	obj			this is a vector with the final objective value for each round

a list consisting of
	alpha		alpha for SVs as vector
	SV		support vectors as matrix
}
\description{
Use lasvm to train a given problem.

Use lasvm to train a given problem.
}

